requirements.txt
langchain
langchain_openai
langchain_groq
python-dotenv
langchain-community
fitz
rank-bm25
PyMuPDF
deepeval
fastembed
pypdf
faiss-cpu
chromadb
langgraph
langsmith==0.1.83
markitdown
python-docx
streamlit
langchain_chroma
langchain_google_vertexai
langchain_google_genai
google-generativeai
streamlit-custom-notification-box
streamlit-option-menu
streamlit-searchbox
unicodedata2
langchain-openai
vertexai
fastapi
python-multipart
uvicorn
httpx
pandas
openpyxl